requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
GeoAlchemy2==0.14.1
python-dotenv==1.0.0
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
